+++ animals/animal.py
@@ -1,3 +1,5 @@
+
+
 class Animal:
 
     def __init__(self, species):
