+++ animals/river_dolphin.py
@@ -1,6 +1,7 @@
 from animals import Animal
 from animals import Freshwater
-from animals import Identifiable
+from identifiable import Identifiable
+
 
 class RiverDolphin(Animal, Freshwater, Identifiable):
 
@@ -8,7 +9,7 @@ class RiverDolphin(Animal, Freshwater, Identifiable):
         Animal.__init__(self, "River dolphin")
         Freshwater.__init__(self)
         Identifiable.__init__(self)
-        self.__prey = { "Trout", "Mackarel", "Salmon", "Sardine" }
+        self.__prey = {"Trout", "Mackarel", "Salmon", "Sardine"}
 
     @property
     def prey(self):
@@ -19,7 +20,6 @@ class RiverDolphin(Animal, Freshwater, Identifiable):
             print(f'The dolphin ate {prey} for a meal')
         else:
             print(f'The dolphin rejects the {prey}')
-
 
     def __str__(self):
         return f'Dolphin {self.id}. Eeee EeeEEeeeeEE!'
