+++ plants/plant.py
@@ -1,5 +1,5 @@
 class Plant:
 
     def __init__(self, species, season):
-      self.species = species
-      self.peak_season = season
+        self.species = species
+        self.peak_season = season
